test: split PDF visual inline display boundary coverage
